LayoutTests/platform/ios/imported/w3c/web-platform-tests/editing/other/typing-space-in-editable-button.tentative-expected.txt

HelloWorld
HelloWorld

HelloWorld

FAIL Type space key in <button contenteditable> should be handled by the <button> assert_not_equals: Click event should be fired on the <button> got disallowed value null
FAIL Type space key in editable <button> shouldn't be handled by the <button> when it's not focused assert_equals: A space should be inserted into the button label expected "Hello World" but got "HelloWorld"
FAIL Type space key in editable <button> should be handled by the <button> when it's focused assert_not_equals: Click event should be fired on the <button> got disallowed value null
FAIL Type space key in editable element in <button> shouldn't be handled by the <button> assert_equals: A space should be inserted into the button label expected "Hello World" but got "HelloWorld"
